Rename queries to gql_queries
