+++ gql_queries.py
